requirements/dev.txt
-r base.txt
autoflake==2.3.1
black==25.1.0
cfgv==3.4.0
click==8.2.1
colorama==0.4.6
distlib==0.4.0
django-debug-toolbar
django-extensions
filelock==3.19.1
identify==2.6.13
iniconfig==2.1.0
isort==6.0.1
model-bakery==1.20.5
mypy_extensions==1.1.0
nodeenv==1.9.1
packaging==25.0
pathspec==0.12.1
platformdirs==4.3.8
pluggy==1.6.0
pre_commit==4.3.0
pyflakes==3.4.0
Pygments==2.19.2
pytest==8.4.2
pytest-django==4.11.1
python-dotenv==1.1.1
pytest-cov
PyYAML==6.0.2
ruff==0.12.10
virtualenv==20.34.0
